requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
pydantic-settings==2.6.1

# Git Operations
GitPython==3.1.41

# AI/LLM APIs
openai==1.10.0
ollama==0.1.6

# HTTP Client (ollama 0.1.6 requires httpx<0.26.0)
httpx==0.25.2

# Retry Logic
tenacity==8.2.3

# Environment Variables
python-dotenv==1.0.0

# Utilities
aiofiles==23.2.1
